requirements.txt
flask
flask-cors
gunicorn
pymupdf
python-pptx
pillow